Refactor and add rawChunk
