+++ src/index.js
@@ -5,3 +5,4 @@ export { Chunk } from "./Chunk.jsx";
 export { ChunkCollection } from "./ChunkCollection.jsx";
 export { ChunkFieldValue } from "./ChunkFieldValue.jsx";
 export { useChunk } from "./useChunk";
+export { rawChunk } from './rawChunk'
