+++ src/utils/caching.js
@@ -1,0 +1,9 @@
+const getCachedData = (id) => {
+  return localStorage.getItem(id);
+}
+
+const storeCache = (id, data) => {
+  localStorage.setItem(id, JSON.stringify(data));
+}
+
+export { getCachedData, storeCache }
